pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
description = ""

readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "kirt08-tokens (>=0.1.0,<0.2.0)",
    "kirt08-contracts (>=0.1.7,<0.2.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"